Fixed delete foutes

+++ app/Http/Controllers/CategoriesController.php
@@ -97,6 +97,13 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        // 記事で使用しているタグの場合
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category cannot be deleted, it has some posts.');
+
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully.');
