+++ app/Http/Controllers/TagsController.php
@@ -97,6 +97,12 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0) {
+            session()->flash('error', 'Tag cannot be deleted, it is associated to some posts.');
+
+            return redirect()->back();
+        }
+
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully.');
